+++ packages/react-native-jigsaw/src/components/DatePicker/DatePicker.tsx
@@ -6,7 +6,7 @@ import dateFormat from "dateformat";
 import { withTheme } from "../../core/theming";
 
 import Portal from "../Portal/Portal";
-import Button from "../Button";
+import Button from "../DeprecatedButton";
 import TextField, { Props as TextFieldProps } from "../TextField";
 import Touchable from "../Touchable";
 import theme from "../../styles/DefaultTheme";
